Separate data socket
