+++ infer_server/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod data_socket;
 pub mod endpoints;
 pub mod nn;
 pub mod protocol;
