+++ infer_server/src/main.rs
@@ -3,15 +3,16 @@ use axum::{
     Extension, Router,
 };
 use env_logger::TimestampPrecision;
-use futures::StreamExt;
-use infer_server::{endpoints::named_stream, nn::UltrafaceModel, protocol::ProtoMsg};
+use infer_server::{
+    data_socket::spawn_data_socket,
+    endpoints::{healthcheck, named_stream},
+    nn::UltrafaceModel,
+};
 use infer_server::{endpoints::recv_named_jpg_streams, pubsub::NamedPubSub};
 use std::{net::SocketAddr, sync::Arc};
-use tokio::net::{TcpListener, TcpStream};
-use tokio_util::codec::{Framed, LengthDelimitedCodec};
 
 #[tokio::main]
-async fn main() {
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
     println!("Let's get started!");
 
     env_logger::builder()
@@ -23,27 +24,12 @@ async fn main() {
     //     .expect("Initialize model");
     let pubsub = Arc::new(NamedPubSub::new());
 
-    let pubsub_ = Arc::clone(&pubsub);
-    let handle = tokio::spawn(async move {
-        let addr = "127.0.0.1:3001";
-        let listener = TcpListener::bind(addr).await?;
-
-        loop {
-            let (socket, _) = listener.accept().await?;
-            let pubsub__ = Arc::clone(&pubsub_);
-            tokio::spawn(async move {
-                handle_incoming(socket, pubsub__).await?;
-                Ok::<_, std::io::Error>(())
-            });
-        }
-
-        Ok::<_, std::io::Error>(())
-    });
+    let handle = spawn_data_socket(pubsub.clone()).await;
 
     let app = Router::new()
         .route("/healthcheck", get(healthcheck))
-        .route("/post_jpgs", post(recv_named_jpg_streams))
         .route("/stream", get(named_stream))
+        .route("/post_jpgs", post(recv_named_jpg_streams))
         .layer(Extension(pubsub));
 
     let addr = SocketAddr::from(([127, 0, 0, 1], 3000));
@@ -52,36 +38,6 @@ async fn main() {
         .serve(app.into_make_service())
         .await
         .unwrap();
-}
-
-async fn healthcheck() -> &'static str {
-    "Healthy"
-}
-
-pub async fn handle_incoming(stream: TcpStream, pubsub: Arc<NamedPubSub>) -> std::io::Result<()> {
-    println!("{}: New connection", stream.peer_addr()?);
-
-    let mut transport = Framed::new(stream, LengthDelimitedCodec::new());
-
-    while let Some(Ok(frame)) = transport.next().await {
-        let data = frame;
-        let proto_msg: ProtoMsg = bincode::deserialize(&data[..]).unwrap();
-        if let ProtoMsg::FrameMsg(frame_msg) = proto_msg {
-            let sender = pubsub.get_sender(&frame_msg.id).await;
-            if let Err(e) = sender.send(frame_msg.data) {
-                println!("Send error");
-            }
-        }
-    }
 
     Ok(())
-}
-
-#[cfg(test)]
-mod test {
-    #[test]
-    fn test_identifying_parts() {
-        let id = b"--blabla\r\n";
-        assert!(id.ends_with("\r\n".as_bytes()));
-    }
 }
